+++ src/types.js
@@ -1156,6 +1156,8 @@ TypeInfo.createByValueInNamespace = function({namespace, value}) {
     name = 'integer';
   } else if (name === "number") {
     name = 'decimal';
+  } else if (value instanceof FP_Date) {
+    name = 'date';
   } else if (value instanceof FP_DateTime) {
     name = 'dateTime';
   } else if (value instanceof FP_Time) {